split util to specific packages (#2)
